# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: TheChickenMan B2B Chicken Marketplace API
  version: 1.0.0
  description: Authoritative RESTful API for TheChickenMan commercial poultry platform.
servers:
  - url: /api/v1
    description: Production & Local API Gateway
paths:
...
  /auth/register:
    post:
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, full_name, phone_number, role]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                full_name: { type: string }
                phone_number: { type: string }
                role: { type: string, enum: [ADMIN, SUPPLIER, BUYER, DRIVER] }
      responses:
        '201':
          description: User created
  /auth/login:
    post:
      summary: Authenticate and obtain JWT access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Authentication successful
  /auth/me:
    get:
      summary: Get currently authenticated user details
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
...
  /admin/dashboard:
    get:
      summary: Admin marketplace performance KPIs (GMV, orders, kg sold)
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Dashboard KPIs }
  /admin/audit-logs:
    get:
      summary: Query immutable audit logs
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Audit logs }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT